+++ internal/errors/transport.go
@@ -45,6 +45,7 @@ type TransportError struct {
 	message string
 	err     error
 	details []interface{}
+	address string
 }
 
 func (t *TransportError) Code() int32 {
@@ -60,6 +61,12 @@ type teOpt func(te *TransportError)
 func WithTEReason(reason TransportErrorCode) teOpt {
 	return func(te *TransportError) {
 		te.Reason = reason
+	}
+}
+
+func WithTEAddress(address string) teOpt {
+	return func(te *TransportError) {
+		te.address = address
 	}
 }
 
@@ -106,6 +113,10 @@ func (t *TransportError) Error() string {
 	if len(t.details) > 0 {
 		b.WriteString(", details: ")
 		b.WriteString(fmt.Sprintf("%v", t.details))
+	}
+	if len(t.address) > 0 {
+		b.WriteString(", address: ")
+		b.WriteString(t.address)
 	}
 	return b.String()
 }
@@ -348,7 +359,7 @@ func IsTransportError(err error, codes ...TransportErrorCode) bool {
 	return false
 }
 
-func MapGRPCError(err error) error {
+func MapGRPCError(err error, opts ...teOpt) error {
 	if err == nil {
 		return nil
 	}
@@ -357,12 +368,16 @@ func MapGRPCError(err error) error {
 		return t
 	}
 	if s, ok := grpcStatus.FromError(err); ok {
-		return &TransportError{
+		te := &TransportError{
 			Reason:  transportErrorCode(s.Code()),
 			message: s.Message(),
 			err:     s.Err(),
 			details: s.Details(),
 		}
+		for _, o := range opts {
+			o(te)
+		}
+		return te
 	}
 	return err
 }