+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Added `retry.WithFastBackoff` and `retry.WithFastBackoff` options
 * Refactored `table.CreateSession` as retry operation with options
 * Moved log level from root of repository to package `log`
+* Added details and address to transport error
 
 ## 3.14.4
 * Implemented auto-removing `conn.Conn` from `conn.Pool` with counting usages of `conn.Conn`